Change build_svg() to require plugin and database connection as parameters
This change is required to allow build_svg() to report status messages
to Trac log. Note that old command line testing code is broken by this
code but I have no idea what the testing is supposed to do because it
depends on files not distributed with the source. This patch only adds
an assert that states this fact.